Allow JWT_AUDIENCES and sub JWT claim to be optional.
Also adds two tests that would fail on previous versions.

Documentation does not say that sub claim is required, nor
does the code validate it beside making sure it is present,
so I removed the validation that sub claim is present, and
modify one test that was failing with this change.

Signed-off-by: Even Thomassen <even.thomassen@noriginmedia.com>