Allow JWTAuth to be given a custom secret key and issuer values. (#12)
This makes it possible to use different issuer and/or secret keys
for different endpoints.

Should be completely backwards-compatible. If a secret or issuer
is not specified, it will use config values JWT_SECRET and
JWT_ISSUER as before, and the requires_token decorator is now
available also on the JWTAuth instance.

Additionally added tests for this new use cases.